add signalR

+++ client/src/app/features/checkout/checkout.component.ts
@@ -22,6 +22,8 @@ import { CheckoutReviewComponent } from './checkout-review/checkout-review.compo
 import { CartService } from '../../core/services/cart.service';
 import { CurrencyPipe } from '@angular/common';
 import { MatProgressSpinner } from '@angular/material/progress-spinner';
+import { OrderToCreate, ShippingAddress } from '../../shared/models/order';
+import { OrderService } from '../../core/services/order.service';
 
 @Component({
   selector: 'app-checkout',
@@ -42,6 +44,7 @@ import { MatProgressSpinner } from '@angular/material/progress-spinner';
 export class CheckoutComponent implements OnInit, OnDestroy {
   private stripeService = inject(StripeService);
   private accountService = inject(AccountService);
+  private orderSerivce = inject(OrderService);
   private snackbar = inject(SnackbarService);
   private router = inject(Router);
   cartService = inject(CartService);
@@ -107,7 +110,7 @@ export class CheckoutComponent implements OnInit, OnDestroy {
   async onStepChange(event: StepperSelectionEvent) {
     if (event.selectedIndex === 1) {
       if (this.saveAddress) {
-        const address = await this.getAddressFromStripeAddress();
+        const address = (await this.getAddressFromStripeAddress()) as Address;
         address && firstValueFrom(this.accountService.updateAddress(address));
       }
     }
@@ -124,12 +127,25 @@ export class CheckoutComponent implements OnInit, OnDestroy {
     try {
       if (this.confirmationToken) {
         const result = await this.stripeService.confirmPayment(this.confirmationToken);
+
         if (result.error) {
           throw new Error(result.error.message);
         }
-        this.cartService.deleteCart();
-        this.cartService.selectedDelivery.set(null);
-        this.router.navigateByUrl('/checkout/success');
+
+        if (result.paymentIntent?.status === 'succeeded') {
+          const order = await this.CreateOrderModal();
+          const orderResult = await firstValueFrom(this.orderSerivce.createOrder(order));
+          if (orderResult) {
+            this.orderSerivce.orderComplete = true;
+            this.cartService.deleteCart();
+            this.cartService.selectedDelivery.set(null);
+            this.router.navigateByUrl('/checkout/success');
+          } else {
+            throw new Error('Order creation failed');
+          }
+        } else {
+          throw new Error('Something went wrong');
+        }
       }
     } catch (error: any) {
       this.snackbar.error(error.message || 'Something went wrong');
@@ -139,12 +155,35 @@ export class CheckoutComponent implements OnInit, OnDestroy {
     }
   }
 
-  private async getAddressFromStripeAddress(): Promise<Address | null> {
+  private async CreateOrderModal(): Promise<OrderToCreate> {
+    const cart = this.cartService.cart();
+    const shippingAddress = (await this.getAddressFromStripeAddress()) as ShippingAddress;
+    const card = this.confirmationToken?.payment_method_preview.card;
+
+    if (!cart?.id || !cart.deliveryMethodId || !shippingAddress || !card) {
+      throw new Error('Problem creating order');
+    }
+
+    return {
+      cartId: cart.id,
+      paymentSummary: {
+        last4: +card.last4,
+        brand: card.brand,
+        expMonth: card.exp_month,
+        expYear: card.exp_year,
+      },
+      deliveryMethodId: cart.deliveryMethodId,
+      shippingAddress,
+    };
+  }
+
+  private async getAddressFromStripeAddress(): Promise<Address | ShippingAddress | null> {
     const result = await this.addressElement?.getValue();
     const address = result?.value.address;
 
     if (address) {
       return {
+        name: result.value.name,
         line1: address.line1,
         line2: address.line2 || undefined,
         city: address.city,
